% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descStat_scores_ranks.R
\name{descStats_scores_ranks}
\alias{descStats_scores_ranks}
\title{Computes average, median, stdev, and coefficient of variation for ranks or scores from multiple simulations.
Returns master file with initial rank (score), followed by average and median one for each statistical unit and each period.}
\usage{
descStats_scores_ranks(x, path, time = c(0), decim = 2)
}
\arguments{
\item{x}{A list containing one or multiple dataframes (one for each year) with the ranks (scores) for different simulations.
The dataframes inside the list have to look like the ones returned by linearAggExc():
Column 1 - statistical units (countries, regions, companies etc.).
Column 2 - time (year, or month, or day etc. - numeric).
Rest of columns - ranks (scores) in simulations.}

\item{path}{Specifies the path to the folder where the returned .csv files ("Master file", "Average", "Median", "Stdev", "CV") are to be saved.}

\item{time}{A vector of the time periods for each dataframe in list x. Default is c(0).}

\item{decim}{Numeric. Specifies the number of decimals to display. Default is 2.}
}
\value{
If path is missing, it returns a list of five data frames (masterFile, x_average, x_median, x_stdev, x_CV).
If path is specified, it also writes five .csv files ("Master file", "Average", "Median", "Stdev", "CV") in the path folder.
Countries are written in column 1. Time is written as colnames.
}
\description{
Computes average, median, stdev, and coefficient of variation for ranks or scores from multiple simulations.
Returns master file with initial rank (score), followed by average and median one for each statistical unit and each period.
}
\examples{
###example for one year
ranks2001[1:5,1:5]
dataTestList=list(ranks2001)
a=descStats_scores_ranks(dataTestList,
#path="E:/Sample folder"
)
a$masterFile
###example for multiple years
names(ranks2001_2003)
a=descStats_scores_ranks(ranks2001_2003, time=c(2001, 2002, 2003), decim=4)
a$masterFile[1:5,]
a$x_stdev[1:5,]
}
