% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearAggExc.R
\name{linearAggExc}
\alias{linearAggExc}
\title{Computes index scores and ranks using linear aggregation for one or multiple time periods.
Automatically computes alternative scores and ranks for inclusion and exclusion of individual indicators.}
\usage{
linearAggExc(x, direction, weight, path, year1, yearT)
}
\arguments{
\item{x}{A dataframe. Column 1 - statistical units (countries, regions, companies etc.).
Column 2 - time (year, or month, or day etc. - numeric).
Rest of columns - values for individual variables.
The pre-processing (e.g. standardization, winsoriztion) has to be done by the researcher before using this function.
For standardization, one may use the rescaleColumns() function of this package.}

\item{direction}{A vector. Specifies the direction (as 1 or -1) for each individual variable in the index.}

\item{weight}{A vector. Specifies the weight assigned to each individual variable.}

\item{path}{Specifies the path to the folder where the returned .csv files of scores and ranks are to be saved.}

\item{year1}{Numeric. Use for multiple years (time periods). Specifies the first year (period) in x.}

\item{yearT}{Numeric. Use for multiple years (time periods). Specifies the last year (period) in x.}
}
\value{
Writes two .csv files per year - Index scores and Ranks - in the path folder.
Both index scores and ranks are computed using all individual variables (column 3)
and when alternatively dropping one individual variable (columns 4 to last).
Countries and time are written in column 1 and 2, respectively.
In case of one year data (x), the function returns two dataframes - scores and ranks.
}
\description{
Computes index scores and ranks using linear aggregation for one or multiple time periods.
Automatically computes alternative scores and ranks for inclusion and exclusion of individual indicators.
}
\examples{
dataTest_oneYear[1:5,1:5]
direction=c(1, -1,	1,	-1,	1,	-1,	1,	-1,	1,	1,	1,	1,	1,	1,	1,	-1,	-1)
weight=c(1/18, 1/18, 1/18, 1/18, 1/18, 1/18, 1/15, 1/15, 1/15, 1/75, 1/75, 1/75, 1/75, 1/75, 1/15, 1/6, 1/6)
a=linearAggExc(dataTest_oneYear, direction, weight)
#to also store the scores and ranks into 2 .csv files:
#a=linearAggExc(dataTest, direction, weight, path="E:/Sample folder")
a$scores[1:5,1:5]
a$ranks[1:5,1:5]
direction=c(1, -1,	1,	-1,	1,	-1,	1,	-1,	1,	1,	1,	1,	1,	1,	1,	-1,	-1)
weight=c(1/18, 1/18, 1/18, 1/18, 1/18, 1/18, 1/15, 1/15, 1/15, 1/75, 1/75, 1/75, 1/75, 1/75, 1/15, 1/6, 1/6)
#linearAggExc(x=dataTest_moreYears, direction=direction, weight=weight, path="E:/Sample folder", year1=2001, yearT=2003)
}
\references{
For theoretical information on the linear aggregation method and on the inclusion-exclusion of individual indicators:
Nardo et al.(2008). Handbook on Constructing Composite Indicators and User Guide. Publisher: OECD; Editor: European Commission, Joint Research Centre and OECD.
}
