% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranksCount.R
\name{ranksCount}
\alias{ranksCount}
\title{Computes number and percent of times that each rank is taken by each statistical unit.}
\usage{
ranksCount(x, decim = 2, path, year = "")
}
\arguments{
\item{x}{A dataframe containing the ranks for different simulations - such as returned by linearAggExc().
Column 1 - statistical units (countries, regions, companies etc.).
Column 2 - time (year, or month, or day etc. - numeric).
Rest of columns - ranks in simulations.}

\item{decim}{Numeric. Specifies the number of decimals to display in case of percent. Default is 2.}

\item{path}{Specifies the path to the folder where the returned .csv files of ranksNb and ranksPc are to be saved.}

\item{year}{Specifies the period in case the user wants the retrned files to be saved as "RanksNb year.csv" and "RanksPc year.csv"}
}
\value{
If path is missing, it returns a list of two data frames (ranksNb and ranksPc).
If path is specified, it also writes two .csv files (RanksNb and RanksPc) in the path folder.
Countries and time are written in column 1 and 2, respectively.
}
\description{
Computes number and percent of times that each rank is taken by each statistical unit.
}
\examples{
ranks2001[1:5,]
a=ranksCount(ranks2001)
a$rankNb[1:5,]
a$rankPc[1:5,]
###to automatically store the two .csv files in the desired folder, use path:
#ranksCount(ranks2001, decim=0, path="E:/Sample folder")
}
