% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranksCount_colCode.R
\name{ranksCount_colCode}
\alias{ranksCount_colCode}
\title{Returns a color coded table with frequency of each rank.
Initial, median and modal rank frequency cells are color-coded.}
\usage{
ranksCount_colCode(x, decim = 2, path, year = "")
}
\arguments{
\item{x}{A dataframe containing the ranks for different simulations - such as returned by linearAggExc().
Column 1 - statistical units (countries, regions, companies etc.).
Column 2 - time (year, or month, or day etc. - numeric).
Rest of columns - ranks in simulations.}

\item{decim}{Numeric. Specifies the number of decimals to display. Default is 2.}

\item{path}{Specifies the path to the folder where the returned .xlsx file ("Color coded ranks frequency.xlsx") to be saved.}

\item{year}{Specifies the period in case the user wants the returned file to be saved as "Color coded ranks frequency year.xlsx".}
}
\value{
A table with: Column 1- statistical units; column 2- time; column 3- frequency of taking rank 1; and so on.
For each statistical unit:
The cell that shows the frequency of the initial rank is bordered in red.
The frequency of the median rank is bolded.
The frequency of the modal rank is written in blue.
If path is specified, it also writes a .xlsx file in the path folder.
The function was created haing in mind the .xlsx export.
}
\description{
Returns a color coded table with frequency of each rank.
Initial, median and modal rank frequency cells are color-coded.
}
\examples{
ranks2001[1:5,1:5]
#ranksCount_colCode(ranks2001, decim=4, path="E:/Sample folder", year="2001")
a=ranksCount_colCode(ranks2001,decim=0)
a
}
