% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescaleColumns.R
\name{rescaleColumns}
\alias{rescaleColumns}
\title{Rescales (minMax, z-score, rank) values per column in a dataframe.}
\usage{
rescaleColumns(x, method = "minMax", max = 100, decim = 2)
}
\arguments{
\item{x}{A dataframe. Column 1 - statistical units (countries, regions, companies etc.).
Column 2 - time (year, or month, or day etc. - numeric).
Rest of columns - values for individual variables.}

\item{method}{Specifies the rescaling method: "minmax", "z-score", or "rank".
The default is "minmax". Equal ranks are averaged.}

\item{max}{Numeric. In case of method="minmax", specifies the value to be taken by the maximum value in a column.
The default is 100.}

\item{decim}{Numeric. Specifies the number of decimals to display. Default is 2.}
}
\value{
A matrix. The first two columns and the colnames are the same as those of x.
}
\description{
Rescales (minMax, z-score, rank) values per column in a dataframe.
}
\examples{
dataTest_moreYears[c(1:5,27:32),1:5]
rescaleColumns(dataTest_moreYears)[c(1:5,27:32),1:5]
rescaleColumns(dataTest_moreYears, method="z-score", decim=3)[1:5,1:5]
}
\references{
For theoretical information on the three rescaling methods, check:
Nardo et al.(2008). Handbook on Constructing Composite Indicators and User Guide. Publisher: OECD; Editor: European Commission, Joint Research Centre and OECD.
}
